remove batch from v0, refactor names, change java stub gens
