+++ grpc/proto/v0/search_get.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 package weaviate.v0;
 
 option go_package = "github.com/weaviate/weaviate/grpc/generated;protocol";
-option java_package = "io.weaviate.grpc.protocol";
+option java_package = "io.weaviate.grpc.protocol.v0";
 option java_outer_classname = "WeaviateProtoSearchGet";
 
 message SearchRequest {
