+++ grpc/proto/v0/weaviate.proto
@@ -6,7 +6,7 @@ import "v0/batch.proto";
 import "v0/search_get.proto";
 
 option go_package = "github.com/weaviate/weaviate/grpc/generated;protocol";
-option java_package = "io.weaviate.grpc.protocol";
+option java_package = "io.weaviate.grpc.protocol.v0";
 option java_outer_classname = "WeaviateProto";
 
 service Weaviate {
