+++ grpc/proto/v1/base.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 package weaviate.v1;
 
 option go_package = "github.com/weaviate/weaviate/grpc/generated;protocol";
-option java_package = "io.weaviate.grpc.protocol";
+option java_package = "io.weaviate.grpc.protocol.v1";
 option java_outer_classname = "WeaviateProtoBase";
 
 enum ConsistencyLevel {
