+++ grpc/proto/v1/weaviate.proto
@@ -6,7 +6,7 @@ import "v1/batch.proto";
 import "v1/search_get.proto";
 
 option go_package = "github.com/weaviate/weaviate/grpc/generated;protocol";
-option java_package = "io.weaviate.grpc.protocol";
+option java_package = "io.weaviate.grpc.protocol.v1";
 option java_outer_classname = "WeaviateProto";
 
 service Weaviate {
